+++ src/main/java/com/library/dto/BookDTO.java
@@ -4,7 +4,7 @@ import java.util.Objects;
 
 public class BookDTO {
 
-    private long id;
+    private int id;
     private String author;
     private String name;
     private int countPages;
@@ -15,7 +15,7 @@ public class BookDTO {
     public BookDTO() {
     }
 
-    public BookDTO(long id, String author, String name, int countPages, int publicationYear, String description, Genre genre) {
+    public BookDTO(int id, String author, String name, int countPages, int publicationYear, String description, Genre genre) {
         this.id = id;
         this.author = author;
         this.name = name;
@@ -33,11 +33,11 @@ public class BookDTO {
         this.author = author;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
